requirements.txt
Django>=4.0,<5.0
psycopg2>=2.8
django-extensions>=3.0,<4.0
djangorestframework>=3.14.0
Pillow>=10.1.0
django-environ==0.9.0